auth

+++ routes/web.php
@@ -1,6 +1,8 @@
 <?php
 
 use App\Http\Controllers\JobController;
+use App\Http\Controllers\RegisteredUserController;
+use App\Http\Controllers\SessionController;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Facades\Route;
 use App\Models\Job;
@@ -54,24 +56,30 @@ Route::get('/jobs/create', [JobController::class, 'create']);
 Route::get('/jobs/{job}', [JobController::class, 'show']);
 
 //Store
-Route::post('/jobs', [JobController::class, 'store']);
+Route::post('/jobs', [JobController::class, 'store'])->middleware('auth');
 
 
 //Edit
-Route::get('/jobs/{job}/edit', [JobController::class, 'edit']);
+//Route::get('/jobs/{job}/edit', [JobController::class, 'edit'])->middleware(['auth', 'can:edit-job,job']);
+Route::get('/jobs/{job}/edit', [JobController::class, 'edit'])->middleware('auth')->can('edit-job', 'job');
 
 
 //Update
-Route::patch('/jobs/{job}', [JobController::class, 'update']);
+Route::patch('/jobs/{job}', [JobController::class, 'update'])->middleware('auth')->can('edit-job', 'job');;
 
 
 //Destroy
-Route::delete('/jobs/{job}', [JobController::class, 'destroy']);
+Route::delete('/jobs/{job}', [JobController::class, 'destroy'])->middleware('auth')->can('edit-job', 'job');;
 
 Route::view('/', 'contacts');
 
+Route::get('/register', [RegisteredUserController::class, 'create']);
+Route::post('/register', [RegisteredUserController::class, 'store']);
 
+Route::get('/login', [SessionController::class, 'create'])->name('login');
 
+Route::post('/login', [SessionController::class, 'store']);
+Route::post('/logout', [SessionController::class, 'destroy']);
 //#3
 //Route::resource('jobs', JobController::class);
